[DEV] readData method is fully operable
Changes to the python notebook reflect the changes made in the readData method in its current form. readData function saves inputs to parameters as a list of either np.array or str depending on what's needed. This is the largest portion of the Material container that was left to develop so next updates will be dedicated to updating documentation and error testing.