[DEV] Control Drum and Stationary Reflector Clearance
Void gap implemented between control drum and stationary reflector. Next step is to add a 1/16" ss316 layering onto drums as images would suggest housing was used, although this was not explicitly mentioned in documentation. Note that the computed k-eff for this model in the C-3 configuration is within the 2% uncertainty tolerance indicated within Table 2 of NAA-SR-9642.