[DEV] Fuel Composition, reflector material mixing, and filling coolant passaged with air
The fuel composition was recomputed after noticing that the uranium loading and hydrogen density cited in NAA-SR-9642 was different than the loading provided in AI-AEC-13070, which was the original document used for fuel composition. The original computed fuel composition was left in as UZrH_Summary. Additionally, due to challenges with understanding dimensioning quantities it was opted to instead use Serpents mix card to create a new material reflMix, which would follow the volume fraction found in Table 2 of NAA-SR-9642 for the internal reflector material. Lastly, based on NAA-SR-9642 although not explicitly stated, there is nothing to suggest the experiment was under vacuum conditions thus it made more sense to include air in the coolant channels rather than void.