[Dev} Shim B cutout for drive shaft
B shims have been redefined to remove material from the center of the shim to allow for inclusion of the drive shaft. This should help reactivity values for C-1, C-2, and C-5 critical configurations. Model for B shims was visually approximated based on Figure 3 of the SNAP 8 Summary document (AI-AEC-13070).